requirements.txt
dify_plugin>=0.4.0,<0.7.0
requests>=2.31.0
Pillow>=10.0.0
